adding 1019

+++ 1805 | [P1] Natural Sum.cpp
@@ -3,15 +3,15 @@ using namespace std;
 int main()
 {
 
-    int a, b;
+    long long int a, b;
     short int i;
-    int sum = 0;
+    long long int sum = 0;
     cin >> a >> b;
 
     sum = (a + b) * (b - a + 1) / 2; // formula if first and last value have
     // for (i = first_value; i <= last_value; i++)
     // {
-    //     sum += i;                                // don't use loop it's take time
+    //     sum += i;               // don't use loop, it's take time
     // }
     cout << sum << endl;
